Separating out service layer to httpUtils
